grafcet-agents/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
google-generativeai>=0.8.6
google-genai>=1.67.0
google-adk>=1.0.0
pydantic>=2.10.0
python-dotenv>=1.0.1
websockets>=14.2
aiohttp>=3.11.0
aiofiles>=24.1.0
pypdf>=5.3.0
pyautogui>=0.9.54
Pillow>=12.1.1
opencv-python-headless>=4.11.0
httpx>=0.28.0
google-cloud-storage>=2.10.0
